requirements.txt
adjustText==0.7.3
altair==4.1.0
appnope==0.1.2
argcomplete==1.12.3
argon2-cffi==21.1.0
astor==0.8.1
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.0
bleach==4.1.0
blinker==1.4
Brotli==1.0.9
cachetools==4.2.4
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.6
click==7.1.2
cycler==0.10.0
dash==2.0.0
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
docopt==0.6.2
entrypoints==0.3
et-xmlfile==1.1.0
ete3==3.1.2
Flask==2.0.1
Flask-Compress==1.10.1
fugle-realtime==0.2.10
ghp-import==2.0.2
gitdb==4.0.9
GitPython==3.1.24
google-auth==2.3.2
gsheetsdb==0.1.13
idna==3.2
importlib-metadata==4.8.1
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jsonschema==4.1.2
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mergedeep==1.3.4
mgt2001==1.0.0
mistune==0.8.4
mkdocs==1.2.3
mkdocs-material==7.3.5
mkdocs-material-extensions==1.0.3
mo-dots==4.22.21108
mo-future==3.147.20327
mo-imports==3.149.20327
mo-kwargs==4.22.21108
mo-logs==4.23.21108
moz-sql-parser==4.40.21126
mplfinance==0.12.7a17
mpmath==1.2.1
multitasking==0.0.9
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
numpy==1.21.2
openpyxl==3.0.9
packaging==21.0
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.21
protobuf==3.19.0
ptyprocess==0.7.0
pyarrow==6.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pydeck==0.7.1
Pygments==2.10.0
pymdown-extensions==9.0
pyparsing==3.0.3
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyyaml-env-tag==0.1
pyzmq==22.3.0
requests==2.26.0
rsa==4.7.2
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
statsmodels==0.13.0
stemgraphic==0.9.1
streamlit==1.1.0
sympy==1.9
tenacity==8.0.1
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.1.1
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.0.2
urllib3==1.26.7
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
yfinance==0.1.64
zipp==3.5.0
